닉네임 랜덤 생성부 presenter로 통합
