+++ app/src/main/java/com/unimind/unithing/Contract/NicknameContract.kt
@@ -1,0 +1,14 @@
+package com.unimind.unithing.Contract
+
+import com.unimind.unithing.Data.User
+
+interface NicknameContract {
+    interface View{
+
+    }
+    interface Presenter{
+        fun makeRandomId(): String
+    }
+    interface NicknameRepository{
+    }
+}
