+++ app/src/main/java/com/unimind/unithing/Presenter/UserInfoPresenter.kt
@@ -14,5 +14,4 @@ class UserInfoPresenter(val view: UserInfoContract.View): UserInfoContract.Prese
             view.isAuthorized()
         } else view.notAuthorized()
     }
-
 }
